Move concrete constraints to their own module.
